7series MGT rework
I can't port '.xci' (or '.dcp') files between different 7-series
devices (e.g., a 7z030 and 7z045 which have the same MGTs)!

Thus, use TCL to generate IPs. A bit of added time to synthesis
for peace of mind...

Also: rename wrapper to TimingGtCoreWrapper (since the entity
for GTP and GTX is the same; TCL loads different architectures
for GTP and GTX.)