EvrV2Trigger, EvrV2CoreTriggers: added inputs to invert polarity
This is useful if the outputs drive differential pins and the p/n
signals are swapped on the PCB for routing reasons.

The inversion could be achieved by an additional register outside
of the EvrV2CoreTriggers module but then an additional delay is
introduced (which may be surprising to the user).

A combinatorial inversion at the output is undesirable if the
user wants to implement the last register in the IOB.